Added the endpoint buildDetailsForBuild

+++ TravisCIWrapper/Framework/TWBuild.h
@@ -16,12 +16,12 @@
 @property (nonatomic, strong) NSDate *finishedAt;
 @property (nonatomic, assign) NSInteger buildId;
 @property (nonatomic, strong) NSString *message;
-@property (nonatomic, assign) NSInteger buildNumber;
+@property (nonatomic, assign) float buildNumber;
 @property (nonatomic, assign) NSInteger repoId;
 @property (nonatomic, assign) NSInteger result;
 @property (nonatomic, strong) NSDate *startedAt;
 @property (nonatomic, strong) NSString *state;
 
-- (id)initWithBranch:(NSString *)branch commit:(NSString *)commit duration:(NSInteger)duration eventType:(NSString *)eventType finishedAt:(NSDate *)finishedAt buildId:(NSInteger)buildId message:(NSString *)message buildNumber:(NSInteger)buildNumber repoId:(NSInteger)repoId result:(NSInteger)result startedAt:(NSDate *)startedAt state:(NSString *)state;
+- (id)initWithBranch:(NSString *)branch commit:(NSString *)commit duration:(NSInteger)duration eventType:(NSString *)eventType finishedAt:(NSDate *)finishedAt buildId:(NSInteger)buildId message:(NSString *)message buildNumber:(float)buildNumber repoId:(NSInteger)repoId result:(NSInteger)result startedAt:(NSDate *)startedAt state:(NSString *)state;
 
 @end
