+++ TravisCIWrapper/Framework/TWRepoDetails.m
@@ -13,7 +13,7 @@
 - (id)initWithRepoDescription:(NSString *)description repoId:(NSInteger)repoId
             lastBuildDuration:(NSInteger)lastBuildDuration lastBuildFinishedAt:(NSDate *)lastBuildFinishedAt
                   lastBuildId:(NSInteger)lastBuildId lastBuildLanguage:(NSString *)lastBuildLanguage
-              lastBuildNumber:(NSInteger)lastBuildNumber lastBuildResult:(NSInteger)lastBuildResult
+              lastBuildNumber:(float)lastBuildNumber lastBuildResult:(NSInteger)lastBuildResult
            lastBuildStartedAt:(NSDate *)lastBuildStartedAt lastBuildStatus:(NSInteger)lastBuildStatus
                          slug:(NSString *)slug publicKey:(NSString *)publicKey {
     self = [self initWithRepoDescription:description repoId:repoId lastBuildDuration:lastBuildDuration lastBuildFinishedAt:lastBuildFinishedAt lastBuildId:lastBuildId lastBuildLanguage:lastBuildLanguage lastBuildNumber:lastBuildNumber lastBuildResult:lastBuildResult lastBuildStartedAt:lastBuildStartedAt lastBuildStatus:lastBuildStatus slug:slug];
